tests: update config tests to match src move
