--- tests/dataio/test_config.py
@@ -1,27 +1,0 @@
-import os
-
-import pydantic as pyd
-import pytest
-
-from student_success_tool import dataio
-
-FIXTURES_PATH = "tests/fixtures"
-
-
-class BadProjectConfig(pyd.BaseModel):
-    is_bad: bool = pyd.Field(...)
-
-
-@pytest.mark.parametrize(
-    ["file_name", "schema", "context"],
-    [
-        # TODO
-        # ("project_config.toml", PDPProjectConfig, does_not_raise()),
-        ("project_config.toml", BadProjectConfig, pytest.raises(pyd.ValidationError)),
-    ],
-)
-def test_load_config(file_name, schema, context):
-    file_path = os.path.join(FIXTURES_PATH, file_name)
-    with context:
-        result = dataio.read_config(file_path, schema=schema)
-        assert isinstance(result, pyd.BaseModel)
